+++ client-config/src/test/java/com/palantir/conjure/java/client/config/ClientConfigurationsTest.java
@@ -70,8 +70,7 @@ public final class ClientConfigurationsTest {
         X509TrustManager trustManager = mock(X509TrustManager.class);
         ClientConfiguration actual = ClientConfigurations.of(uris, sslFactory, trustManager);
 
-        assertThat(actual.sslSocketFactory()).isInstanceOf(KeepAliveSslSocketFactory.class);
-        assertThat(((KeepAliveSslSocketFactory) actual.sslSocketFactory()).getDelegate()).isEqualTo(sslFactory);
+        assertThat(actual.sslSocketFactory()).isEqualTo(sslFactory);
         assertThat(actual.trustManager()).isEqualTo(trustManager);
         assertThat(actual.uris()).isEqualTo(uris);
         assertThat(actual.connectTimeout()).isEqualTo(Duration.ofSeconds(10));
