+++ okhttp-clients/src/main/java/com/palantir/conjure/java/okhttp/KeepAliveSslSocketFactory.java
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2020 Palantir Technologies Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.palantir.conjure.java.client.config;
+package com.palantir.conjure.java.okhttp;
 
 import com.palantir.logsafe.Preconditions;
 import java.net.Socket;
